+++ flutter_value_bloc/lib/src/internalUtils.dart
@@ -1,14 +1,16 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class ViewCubitInitializer<C extends Cubit<Object>> extends StatefulWidget {
+class ViewCubitInitializer<C extends Cubit<S>, S> extends StatefulWidget {
   final C cubit;
+  final bool Function(S state) initializeWhen;
   final void Function(BuildContext context, C cubit) initializer;
   final Widget child;
 
   const ViewCubitInitializer({
     Key key,
     @required this.cubit,
+    @required this.initializeWhen,
     @required this.initializer,
     @required this.child,
   })  : assert(cubit != null),
@@ -17,24 +19,37 @@ class ViewCubitInitializer<C extends Cubit<Object>> extends StatefulWidget {
         super(key: key);
 
   @override
-  _ViewCubitInitializerState<C> createState() => _ViewCubitInitializerState();
+  _ViewCubitInitializerState<C, S> createState() => _ViewCubitInitializerState();
 }
 
-class _ViewCubitInitializerState<C extends Cubit<Object>> extends State<ViewCubitInitializer<C>> {
+class _ViewCubitInitializerState<C extends Cubit<S>, S> extends State<ViewCubitInitializer<C, S>> {
   @override
   void initState() {
     super.initState();
-    widget.initializer(context, widget.cubit);
+    initialize();
   }
 
   @override
-  void didUpdateWidget(covariant ViewCubitInitializer<C> oldWidget) {
+  void didUpdateWidget(covariant ViewCubitInitializer<C, S> oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (widget.cubit != oldWidget.cubit) {
+      initialize();
+    }
+  }
+
+  void initialize() {
+    if (widget.initializeWhen(widget.cubit.state)) {
       widget.initializer(context, widget.cubit);
     }
   }
 
   @override
-  Widget build(BuildContext context) => widget.child;
+  Widget build(BuildContext context) {
+    return BlocListener<C, S>(
+      cubit: widget.cubit,
+      listenWhen: (p, c) => p.runtimeType != c.runtimeType,
+      listener: (context, state) => initialize(),
+      child: widget.child,
+    );
+  }
 }