+++ flutter_value_bloc/lib/src/object/ViewCubitBuilder.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_value_bloc/src/cubit_views/CubitViews.dart';
 import 'package:flutter_value_bloc/src/cubit_views/ValueViewBuilder.dart';
+import 'package:flutter_value_bloc/src/internalUtils.dart';
 import 'package:flutter_value_bloc/src/utils.dart';
 import 'package:value_bloc/value_bloc.dart';
 
@@ -33,7 +34,9 @@ class ViewCubitBuilder<Value> extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<ObjectCubit<Value, Object>, ObjectCubitState<Value, Object>>(
+    final objectCubit = this.objectCubit;
+
+    final current = BlocBuilder<ObjectCubit<Value, Object>, ObjectCubitState<Value, Object>>(
       cubit: objectCubit,
       builder: (context, state) {
         if (state is ObjectCubitUpdating<Value, Object>) {
@@ -53,5 +56,17 @@ class ViewCubitBuilder<Value> extends StatelessWidget {
         return builder(context, state.value);
       },
     );
+
+    if (objectCubit is SingleCubit<Value, Object, Object>) {
+      return ViewCubitInitializer<SingleCubit<Value, Object, Object>,
+          ObjectCubitState<Value, Object>>(
+        cubit: objectCubit,
+        initializeWhen: (state) => state is ObjectCubitUpdating<Value, Object>,
+        initializer: (context, c) => c.fetch(),
+        child: current,
+      );
+    }
+
+    return current;
   }
 }