+++ flutter_value_bloc/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.0
+- Now you can build your ui with the new modular cubits! `ModularViewCubitBuilder`
+- Introduced new builders: `ListViewCubitBuilder`, `GridViewCubitBuilder`
+
 ## 0.4.3
 Added `analysis_options.yaml`