+++ value_bloc/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.5.0
+- Now the cubits are modular inside your cubit! `ValueCubit`, `ListCubit`, `SetCubit`, `SingleCubit` and `MultiCubit`
+- Now you can mix your cubit with loading cubit!
+
 ## 0.4.3
 Added `analysis_options.yaml`